connection retry, UI structure tweaks
