+++ web/index.html
@@ -7,7 +7,6 @@
     <title>DIMO B2B Fleet Manager</title>
     <link rel="stylesheet" href="./monospace-web/index.css" />
     <link rel="stylesheet" href="./monospace-web/reset.css" />
-    <link rel="stylesheet" href="./monospace-web/reset.css" />
     <script type="module" src="./src/index.ts"></script>
   </head>
   <body>
