+++ web/src/elements/add-vin-element.ts
@@ -1,67 +1,10 @@
-import {html, LitElement} from 'lit'
+import {html} from 'lit'
 import {SettingsService} from "@services/settings-service";
 import {customElement, property, state} from "lit/decorators.js";
-import {ApiService} from "@services/api-service.ts";
-import {SigningService} from "@services/signing-service.ts";
 import {repeat} from "lit/directives/repeat.js";
 import './session-timer';
-import qs from 'qs';
 import {range} from "lodash";
-import {delay} from "@utils/utils";
-
-interface VehicleLookup {
-    vin: string;
-    userDeviceId: string;
-    definitionId: string;
-    vehicleTokenId: number;
-    syntheticDeviceTokenId: number;
-}
-
-// interface VehicleOracleLookup {
-//     vehicle: Vehicle;
-// }
-
-// interface Vehicle {
-//     vin: string;
-//     id: string;
-//     tokenId: number;
-//     mintedAt: string;
-//     owner: string;
-//     definition: Definition;
-//     syntheticDevice: SyntheticDevice;
-// }
-
-// interface Definition {
-//     id: string;
-//     make: string;
-//     model: string;
-//     year: number;
-// }
-// interface SyntheticDevice {
-//     id: string;
-//     tokenId: number;
-//     mintedAt: string;
-// }
-
-interface VinOnboardingStatus {
-    vin: string;
-    status: string;
-    details: string;
-}
-
-interface VinsOnboardingResult {
-    statuses: VinOnboardingStatus[];
-}
-
-interface VinMintData {
-    vin: string;
-    typedData: any;
-    signature?: string;
-}
-
-interface VinsMintDataResult {
-    vinMintingData: VinMintData[];
-}
+import {BaseOnboardingElement, SacdInput} from "@elements/base-onboarding-element.ts";
 
 enum Permission {
     NONLOCATION_TELEMETRY= 1,
@@ -76,13 +19,6 @@ enum Permission {
 }
 
 type Permissions = Record<Permission, boolean>
-
-interface SacdInput {
-    grantee: `0x${string}`;
-    permissions: BigInt;
-    expiration: BigInt;
-    source: string
-}
 
 const PERMISSIONS_MAP: Record<number, string> = {
     [Permission.APPROXIMATE_LOCATION]: "APPROXIMATE_LOCATION",
@@ -125,26 +61,15 @@ const defaultPermissions = {
 }
 
 @customElement('add-vin-element')
-export class AddVinElement extends LitElement {
+export class AddVinElement extends BaseOnboardingElement {
     @property({attribute: false})
     private vinsBulk: string | null;
-
-    @property({attribute: false})
-    private processing: boolean;
-
-    @property({attribute: false})
-    private processingMessage: string;
-
-    @property({attribute: false})
 
     @property({attribute: false})
     private email: string | null;
 
     @property({attribute: false})
     private alertText: string;
-
-    @property({attribute: false})
-    private onboardResult: VinOnboardingStatus[];
 
     @property({attribute: false})
     private enableSacd: boolean;
@@ -158,26 +83,18 @@ export class AddVinElement extends LitElement {
     @state() sessionExpiresIn: number = 0;
 
     private settings: SettingsService;
-    private api: ApiService;
 
-    // @ts-ignore
-    private signingService: SigningService;
+
     constructor() {
         super();
         this.vinsBulk = "";
-        this.processing = false;
-        this.processingMessage = "";
         this.email = localStorage.getItem("email");
         this.settings = SettingsService.getInstance();
-        this.api = ApiService.getInstance();
         this.alertText = "";
-        this.signingService = SigningService.getInstance();
 
         this.enableSacd = false;
         this.sacdGrantee = "";
         this.sacdPermissions = defaultPermissions;
-
-        this.onboardResult = []
     }
 
     // Disable shadow DOM to allow inherit css
@@ -278,28 +195,8 @@ export class AddVinElement extends LitElement {
     }
 
     displayFailure(alertText: string) {
-        this.processing = false;
-        this.processingMessage = "";
+        super.displayFailure(alertText);
         this.alertText = alertText;
-    }
-
-    updateResult(result : VinsOnboardingResult) {
-        const statusesByVin: Record<string, VinOnboardingStatus> = {}
-        for (const item of result.statuses) {
-            statusesByVin[item.vin] = item
-        }
-
-        const newResult: VinOnboardingStatus[] = [];
-
-        for (const item of this.onboardResult) {
-            newResult.push({
-                vin: item.vin,
-                status: statusesByVin[item.vin]?.status || "Unknown",
-                details: statusesByVin[item.vin]?.details || "Unknown"
-            })
-        }
-
-        this.onboardResult = newResult
     }
 
     async _submitVINs(_event: MouseEvent) {
@@ -359,178 +256,5 @@ export class AddVinElement extends LitElement {
 
 
         this.processing = false;
-    }
-
-    async verifyVehicles(vins: string[]) {
-        const payload = {
-            vins: vins.map(v => ({vin: v, countryCode: 'USA'}))
-        }
-
-        const submitStatus = await this.api.callApi('POST', '/vehicle/verify', payload, true);
-        if (!submitStatus.success) {
-            return false;
-        }
-
-        let success = true
-        for (const attempt of range(10)) {
-            success = true
-            const query = qs.stringify({vins: vins.join(',')}, {arrayFormat: 'comma'});
-            const status = await this.api.callApi<VinsOnboardingResult>('GET', `/vehicle/verify?${query}`, null, true);
-
-            if (!status.success || !status.data) {
-                return false;
-            }
-
-            for (const s of status.data.statuses) {
-                if (s.status !== 'Success') {
-                    success = false;
-                    break;
-                }
-            }
-
-            this.updateResult(status.data)
-
-            if (success) {
-                break;
-            }
-
-            if (attempt < 9) {
-                await delay(5000);
-            }
-        }
-
-        return success;
-    }
-
-    async getMintingData(vins: string[]) {
-        const query = qs.stringify({vins: vins.join(',')}, {arrayFormat: 'comma'});
-        const mintData = await this.api.callApi<VinsMintDataResult>('GET', `/vehicle/mint?${query}`, null, true);
-        if (!mintData.success || !mintData.data) {
-            return [];
-        }
-
-        return mintData.data.vinMintingData;
-    }
-
-    async signMintingData(mintingData: VinMintData[]) {
-        const result: VinMintData[] = [];
-        for (const d of mintingData) {
-            const signature = await this.signingService.signTypedData(d.typedData);
-
-            if (!signature.success || !signature.signature) {
-                continue
-            }
-
-            result.push({
-                ...d,
-                signature: signature.signature
-            })
-        }
-
-        return result;
-    }
-
-    async submitMintingData(mintingData: VinMintData[], sacd: SacdInput | null) {
-        const payload: {vinMintingData: VinMintData[], sacd?: SacdInput} = {
-            vinMintingData: mintingData,
-        }
-
-        if (sacd !== null) {
-            payload.sacd = sacd
-        }
-
-        const mintResponse = await this.api.callApi('POST', '/vehicle/mint', payload, true);
-        if (!mintResponse.success || !mintResponse.data) {
-            return false;
-        }
-
-        let success = true
-        for (const attempt of range(30)) {
-            success = true
-            const query = qs.stringify({vins: mintingData.map(m => m.vin).join(',')}, {arrayFormat: 'comma'});
-            const status = await this.api.callApi<VinsOnboardingResult>('GET', `/vehicle/mint/status?${query}`, null, true);
-
-            if (!status.success || !status.data) {
-                return false;
-            }
-
-            for (const s of status.data.statuses) {
-                if (s.status !== 'Success') {
-                    success = false;
-                    break;
-                }
-            }
-
-            this.updateResult(status.data)
-
-            if (success) {
-                break;
-            }
-
-            if (attempt < 19) {
-                await delay(5000);
-            }
-        }
-
-        return success;
-    }
-
-    async onboardVINs(vins: string[], sacd: SacdInput | null): Promise<boolean> {
-        let allVinsValid = true;
-        for (const vin of vins) {
-            const validVin = vin?.length === 17
-            allVinsValid = allVinsValid && validVin
-            this.onboardResult.push({
-                vin: vin,
-                status: "Unknown",
-                details: validVin ? "Valid VIN" : "Invalid VIN"
-            })
-        }
-
-        if (!allVinsValid) {
-            this.displayFailure("Some of the VINs are not valid");
-            return false;
-        }
-
-        const verified = await this.verifyVehicles(vins);
-        if (!verified) {
-            this.displayFailure("Failed to verify at least one VIN");
-            return false
-        }
-
-        const mintData = await this.getMintingData(vins);
-        if (mintData.length === 0) {
-            this.displayFailure("Failed to fetch minting data");
-            return false
-        }
-
-        const signedMintData = await this.signMintingData(mintData);
-        const minted = await this.submitMintingData(signedMintData, sacd);
-
-        if (!minted) {
-            this.displayFailure("Failed to onboard at least one VIN");
-            return false;
-        }
-
-        return true
-    }
-    // todo is this deprecated?
-    async registerInOracle(vin :string) {
-        if (vin) {
-            const lookup = await this.getDeviceAPILookup(vin);
-            if (lookup.success && lookup.data) {
-                const {vin, vehicleTokenId} = lookup.data;
-
-                const url = "/vehicle/register";
-                const body = {vin, token_id: vehicleTokenId};
-                await this.api.callApi<any>('POST', url, body, true);
-            }
-        }
-
-    }
-// todo is this deprecated?
-    async getDeviceAPILookup(vin: string) {
-        const url = "/v1/compass/device-by-vin/" + vin;
-        return await this.api.callApi<VehicleLookup>('GET', url, null, true);
     }
 }
